fmt + clippy

+++ src/capture/wgc/d3d.rs
@@ -25,7 +25,7 @@ fn create_d3d_device_with_type(
             None,
             flags,
             None,
-            D3D11_SDK_VERSION as u32,
+            D3D11_SDK_VERSION,
             Some(device),
             None,
             None,
